# src/shop/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация для API интернет-магазина",
        "title": "Shop API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/product": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Создать товар",
                "parameters": [
                    {
                        "description": "Создаваемый товар",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ProductCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получить товар по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Error404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Удалить товар",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Error404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Уменьшить количество товара",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Количество для вычитания",
                        "name": "count",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Error404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получить список товаров",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ProductResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProductsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.Error400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "status": {
                    "type": "string",
                    "example": "invalid JSON"
                }
            }
        },
        "dto.Error404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "status": {
                    "type": "string",
                    "example": "not found"
                }
            }
        },
        "dto.Error500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "status": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
...
        "dto.ProductCreateRequest": {
            "type": "object",
            "required": [
                "available_stock",
                "category",
                "name",
                "price",
                "suppler_id"
            ],
            "properties": {
                "available_stock": {
                    "type": "integer",
                    "example": 120
                },
                "category": {
                    "type": "string",
                    "example": "Alchemy"
                },
                "name": {
                    "type": "string",
                    "example": "Potion of Healing"
                },
                "price": {
                    "type": "number",
                    "example": 49.99
                },
                "suppler_id": {
                    "type": "string",
                    "example": "supplier-abc-123"
                }
            }
        },
        "dto.ProductResponse": {
            "type": "object",
            "properties": {
                "available_stock": {
                    "type": "integer",
                    "example": 120
                },
                "category": {
                    "type": "string",
                    "example": "Alchemy"
                },
                "id": {
                    "type": "string",
                    "example": "product-xyz-789"
                },
                "image_id": {
                    "type": "string",
                    "example": "img-00112233"
                },
                "last_update_date": {
                    "type": "string",
                    "example": "2025-07-01T15:04:05Z"
                },
                "name": {
                    "type": "string",
                    "example": "Potion of Healing"
                },
                "price": {
                    "type": "number",
                    "example": 49.99
                },
                "suppler_id": {
                    "type": "string",
                    "example": "supplier-abc-123"
                }
            }
        },
        "dto.ProductsResponse": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProductResponse"
                    }
                }
            }
        },
...
    }
}